schedulo-backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
motor
pydantic[email]
passlib[bcrypt]
python-jose[cryptography]
python-multipart
bcrypt
